+++ deeplearning/ml4pl/graphs/labelled/dataflow/polyhedra/BUILD
@@ -23,8 +23,8 @@ py_library(
     deps = [
         "//compilers/llvm:opt",
         "//compilers/llvm:opt_util",
-        "//deeplearning/ml4pl/graphs/unlabelled/llvm2graph:graph_builder",
-        "//deeplearning/ml4pl/graphs/unlabelled/llvm2graph/cfg:llvm_util",
+        "//deeplearning/ml4pl/graphs/llvm2graph/legacy:graph_builder",
+        "//deeplearning/ml4pl/graphs/llvm2graph/legacy/cfg:llvm_util",
         "//labm8/py:app",
         "//labm8/py:decorators",
     ],
@@ -37,7 +37,7 @@ py_test(
         ":polyhedra",
         "//compilers/llvm:clang",
         "//compilers/llvm:opt",
-        "//deeplearning/ml4pl/graphs/unlabelled/llvm2graph:graph_builder",
+        "//deeplearning/ml4pl/graphs/llvm2graph/legacy:graph_builder",
         "//labm8/py:test",
         "//third_party/py/networkx",
     ],
