+++ deeplearning/ml4pl/graphs/unlabelled/make_unlabelled_graph_dataset.py
@@ -24,7 +24,7 @@ import typing
 from deeplearning.ml4pl.bytecode import bytecode_database
 from deeplearning.ml4pl.graphs import database_exporters
 from deeplearning.ml4pl.graphs import graph_database
-from deeplearning.ml4pl.graphs.unlabelled.llvm2graph import graph_builder
+from deeplearning.ml4pl.graphs.llvm2graph.legacy import graph_builder
 from labm8.py import app
 from labm8.py import prof
