+++ deeplearning/ml4pl/ir/create/BUILD
@@ -51,8 +51,8 @@ py_library(
         "//compilers/llvm:opt",
         "//compilers/llvm:opt_util",
         "//datasets/linux",
-        "//deeplearning/ml4pl/graphs/unlabelled/llvm2graph/cfg:control_flow_graph",
-        "//deeplearning/ml4pl/graphs/unlabelled/llvm2graph/cfg:llvm_util",
+        "//deeplearning/ml4pl/graphs/llvm2graph/legacy/cfg:control_flow_graph",
+        "//deeplearning/ml4pl/graphs/llvm2graph/legacy/cfg:llvm_util",
         "//deeplearning/ml4pl/ir:ir_database",
         "//labm8/py:app",
         "//labm8/py:decorators",
@@ -96,8 +96,8 @@ py_library(
         "//compilers/llvm:opt_util",
         "//datasets/opencl/device_mapping:opencl_device_mapping_dataset",
         "//deeplearning/clgen/preprocessors:opencl",
-        "//deeplearning/ml4pl/graphs/unlabelled/llvm2graph/cfg:control_flow_graph",
-        "//deeplearning/ml4pl/graphs/unlabelled/llvm2graph/cfg:llvm_util",
+        "//deeplearning/ml4pl/graphs/llvm2graph/legacy/cfg:control_flow_graph",
+        "//deeplearning/ml4pl/graphs/llvm2graph/legacy/cfg:llvm_util",
         "//deeplearning/ml4pl/ir:ir_database",
         "//labm8/py:app",
         "//labm8/py:decorators",
