+++ deeplearning/ml4pl/ir/create/import_from_linux.py
@@ -28,10 +28,10 @@ from compilers.llvm import opt_util
 from datasets.linux import linux
 from deeplearning.ml4pl import ml4pl_pb2
 from deeplearning.ml4pl.bytecode import bytecode_database
-from deeplearning.ml4pl.graphs.unlabelled.llvm2graph.cfg import (
+from deeplearning.ml4pl.graphs.llvm2graph.legacy.cfg import (
   control_flow_graph as cfg,
 )
-from deeplearning.ml4pl.graphs.unlabelled.llvm2graph.cfg import llvm_util
+from deeplearning.ml4pl.graphs.llvm2graph.legacy.cfg import llvm_util
 from labm8.py import app
 from labm8.py import decorators
